+++ src/ForceFactoryRule.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Ebln\PHPStan\EnforceFactory;
 
+use Ebln\Attrib\ForceFactory;
 use PhpParser\Node;
 use PHPStan\Analyser\Scope;
+use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Rules\Rule;
 use PHPStan\Rules\RuleErrorBuilder;
 
@@ -14,6 +16,13 @@ use PHPStan\Rules\RuleErrorBuilder;
  */
 class ForceFactoryRule implements Rule
 {
+    private ReflectionProvider $reflectionProvider;
+
+    public function __construct(ReflectionProvider $reflectionProvider)
+    {
+        $this->reflectionProvider = $reflectionProvider;
+    }
+
     public function getNodeType(): string
     {
         return \PhpParser\Node\Expr\New_::class;
@@ -26,17 +35,18 @@ class ForceFactoryRule implements Rule
             if (!$isName) {
                 continue; // newly instantiated class name couldn't be infered
             }
+            /** @var class-string $class → sadly Psalm cannot be convinced to return class-string for getClassNames in reasonable amount of time */
             $allowedFactories = $this->getAllowedFactories($class);
             if (null === $allowedFactories) {
-                continue; // newly instantiated class dowsn't implement ForceFactory interface
+                continue; // newly instantiated class doesn't implement ForceFactory interface
             }
 
-            if (empty($allowedFactories)) {
+            if ([] === $allowedFactories) {
                 $errors[] = RuleErrorBuilder::message(
-                    ltrim($class, '\\') . ' cannot be instantiated by other classes; see ' . ForceFactoryInterface::class
+                    ltrim($class, '\\') . ' has either no factories defined or a conflict between interface and attribute!'
                 )->build();
 
-                continue;
+                continue; // bogus configuration
             }
 
             /** @psalm-suppress PossiblyNullReference | sad that even phpstan cannot infer that from isInClass */
@@ -57,17 +67,78 @@ class ForceFactoryRule implements Rule
     }
 
     /**
+     * @phpstan-param class-string $className
+     *
      * @return null|string[] List of FQCNs
      *
      * @phpstan-return null|class-string[]
      */
     private function getAllowedFactories(string $className): ?array
     {
-        if (!is_a($className, ForceFactoryInterface::class, true)) {
+        $allowedFactories = $this->getFactoriesFromAttribute($className);
+        if (is_a($className, ForceFactoryInterface::class, true)) {
+            /* phpstan-var class-string<ForceFactoryInterface> $className */
+            $interfaceFactories = $className::getFactories();
+            sort($interfaceFactories);
+            if (null === $allowedFactories) {
+                $allowedFactories = $interfaceFactories;
+            } elseif ($allowedFactories !== $interfaceFactories) {
+                $allowedFactories = []; // Will result in a bogus definition error
+            }
+        }
+
+        return $allowedFactories;
+    }
+
+    /**
+     * @phpstan-param class-string $className
+     *
+     * @return array<class-string>
+     */
+    private function getFactoriesFromAttribute(string $className): ?array
+    {
+        if (\PHP_VERSION_ID < 80000 && $this->reflectionProvider->hasClass($className)) {
             return null;
         }
 
-        return $className::getFactories();
+        $reflection = $this->reflectionProvider->getClass($className);
+        /* psalm-suppress UndefinedClass */
+        $allowedFactories = [];
+        do {
+            /** @psalm-suppress UndefinedClass */
+            $allowedFactories = [...$allowedFactories, ...$this->getFactoriesFromAttributeByClass($reflection->getNativeReflection())];
+        } while ($reflection = $reflection->getParentClass());
+
+        if (empty($allowedFactories)) {
+            return null;
+        }
+        $allowedFactories = array_filter($allowedFactories);
+        sort($allowedFactories);
+
+        return $allowedFactories;
+    }
+
+    /**
+     * @psalm-suppress UndefinedDocblockClass,MismatchingDocblockParamType
+     *
+     * @psalm-param \PHPStan\BetterReflection\Reflection\Adapter\ReflectionClass|\PHPStan\BetterReflection\Reflection\Adapter\ReflectionEnum $reflection
+     *
+     * @return array<int, null|class-string>
+     */
+    private function getFactoriesFromAttributeByClass(\ReflectionClass $reflection): array
+    {
+        /** @psalm-suppress UndefinedClass */
+        foreach ($reflection->getAttributes() as $attribute) {
+            if (ForceFactory::class === $attribute->getName()) {
+                /** @var ForceFactory $forceFactory */
+                $forceFactory     = $attribute->newInstance();
+                $allowedFactories = $forceFactory->getAllowedFactories();
+
+                return empty($allowedFactories) ? [null] : $allowedFactories;
+            }
+        }
+
+        return [];
     }
 
     /**
@@ -81,8 +152,8 @@ class ForceFactoryRule implements Rule
      *
      * @psalm-return  array<array{string, bool}>
      *
-     * @license https://github.com/phpstan/phpstan/blob/1.1.2/LICENSE
-     * @author  Ondřej Mirtes et al. https://github.com/phpstan/phpstan-src/blob/0.12.x/src/Rules/Classes/InstantiationRule.php#blob_contributors_box
+     * @license https://github.com/phpstan/phpstan-src/blob/1.11.x/LICENSE
+     * @author  Ondřej Mirtes et al. https://github.com/phpstan/phpstan-src/blame/1.11.x/src/Rules/Classes/InstantiationRule.php
      * @author  ebln
      *
      * @see     \PHPStan\Rules\Classes\InstantiationRule::getClassNames
@@ -93,8 +164,8 @@ class ForceFactoryRule implements Rule
             return [[(string)$node->class, \true]];
         }
         if ($node->class instanceof \PhpParser\Node\Stmt\Class_) {
-            $anonymousClassType = $scope->getType($node);
-            if (!$anonymousClassType instanceof \PHPStan\Type\TypeWithClassName) {
+            $classNames = $scope->getType($node)->getObjectClassNames();
+            if ([] === $classNames) {
                 throw new \PHPStan\ShouldNotHappenException();
             }
             // Report back extended class!
@@ -102,8 +173,10 @@ class ForceFactoryRule implements Rule
                 return [[$node->class->extends->toString(), \true]];
             }
 
-            // we don't care about the anonymous class' name and abort processing early
-            return [[$anonymousClassType->getClassName(), \false]];
+            return array_map(
+                static fn (string $className) => [$className, \false],
+                $classNames,
+            );
         }
         $type = $scope->getType($node->class);