+++ tests/dataAttrib/IndependentFactory.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Test\Ebln\PHPStan\EnforceFactory\dataAttrib;
+
+use Test\Ebln\PHPStan\EnforceFactory\dataAttrib\code\IndependentForcedFactoryProduct;
+
+class IndependentFactory
+{
+    public function independentClass(): IndependentForcedFactoryProduct
+    {
+        return new IndependentForcedFactoryProduct();
+    }
+}