+++ tests/dataAttrib/code/EmptyProduct.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Test\Ebln\PHPStan\EnforceFactory\dataAttrib\code;
+
+use Ebln\Attrib\ForceFactory;
+
+#[ForceFactory()]
+class EmptyProduct
+{
+}