+++ tests/dataAttrib/code/ExtendedProduct.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Test\Ebln\PHPStan\EnforceFactory\dataAttrib\code;
+
+class ExtendedProduct extends ForcedFactoryProduct
+{
+
+}