+++ tests/dataMixed/code/AttributeProduct.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Test\Ebln\PHPStan\EnforceFactory\dataMixed\code;
+
+use Ebln\Attrib\ForceFactory;
+use Test\Ebln\PHPStan\EnforceFactory\dataMixed\AttributeFactory;
+
+#[ForceFactory(AttributeFactory::class)]
+class AttributeProduct
+{
+}