+++ CHANGELOG.md
@@ -5,11 +5,29 @@ Intended to follow [«Keep a Changelog»](https://keepachangelog.com/en/)
 
 ----
 
-## Upcoming
+## Upcomming
 
-- Support for attributes
+- Remove support for the interface
+- Deprecate (abandon) the interface package
+- create conflict with interface for version 2
+	```json
+	{
+		"conflict": {
+			"ebln/ebln/phpstan-factory-mark": "*"
+		}
+	}
+	```
 
 ----
+
+## [1.0.0]
+
+### Added
+- Support for attributes
+
+### Removed
+* Support for PHP < 7.4
+* Support for PHPStan < 1.11
 
 ## [0.0.2]